+++ pmd-core/src/main/java/net/sourceforge/pmd/RuleSetLoader.java
@@ -17,7 +17,6 @@ import java.util.Properties;
 
 import org.apache.commons.lang3.StringUtils;
 import org.checkerframework.checker.nullness.qual.NonNull;
-import org.checkerframework.checker.nullness.qual.Nullable;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -153,10 +152,6 @@ public final class RuleSetLoader {
         );
     }
 
-    private @Nullable MessageReporter filteredReporter() {
-        return warnDeprecated ? reporter : null;
-    }
-
     /**
      * Parses and returns a ruleset from its location. The location may
      * be a file system path, or a resource path (see {@link #loadResourcesWith(ClassLoader)}).
@@ -166,7 +161,7 @@ public final class RuleSetLoader {
      * @throws RuleSetLoadException If any error occurs (eg, invalid syntax, or resource not found)
      */
     public RuleSet loadFromResource(String rulesetPath) {
-        return loadFromResource(new RuleSetReferenceId(rulesetPath, null, filteredReporter()));
+        return loadFromResource(new RuleSetReferenceId(rulesetPath, null));
     }
 
     /**
@@ -178,12 +173,25 @@ public final class RuleSetLoader {
      * @throws RuleSetLoadException If any error occurs (eg, invalid syntax)
      */
     public RuleSet loadFromString(String filename, final String rulesetXmlContent) {
-        return loadFromResource(new RuleSetReferenceId(filename, null, filteredReporter()) {
-            @Override
-            public InputStream getInputStream(ResourceLoader rl) {
-                return new ByteArrayInputStream(rulesetXmlContent.getBytes(StandardCharsets.UTF_8));
-            }
-        });
+        if (filename == null || filename.isEmpty()) {
+            throw new IllegalArgumentException("Invalid empty filename");
+        }
+
+        ResourceLoader oldLoader = this.resourceLoader;
+        try {
+            loadResourcesWith(new ResourceLoader() {
+                @Override
+                public @NonNull InputStream loadResourceAsStream(String name) throws IOException {
+                    if (Objects.equals(name, filename)) {
+                        return new ByteArrayInputStream(rulesetXmlContent.getBytes(StandardCharsets.UTF_8));
+                    }
+                    return oldLoader.loadResourceAsStream(name);
+                }
+            });
+            return loadFromResource(new RuleSetReferenceId(filename, null));
+        } finally {
+            loadResourcesWith(oldLoader);
+        }
     }
 
     /**