+++ pmd-core/src/main/java/net/sourceforge/pmd/util/ResourceLoader.java
@@ -56,9 +56,9 @@ public class ResourceLoader {
     }
 
     /**
-     * Attempts to load the resource from file, a URL or the claspath
-     * <p>
-     * Caller is responsible for closing the {@link InputStream}.
+     * Attempts to load the resource from file, a URL or the classpath.
+     *
+     * <p>Caller is responsible for closing the {@link InputStream}.
      *
      * @param name The resource to attempt and load
      *
@@ -77,7 +77,7 @@ public class ResourceLoader {
             }
         }
 
-        // Maybe it's a url?
+        // Maybe it's a URL?
         try {
             final HttpURLConnection connection = (HttpURLConnection) new URL(name).openConnection();
             connection.setConnectTimeout(TIMEOUT);